requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
astunparse==1.6.3
attrs==26.1.0
Authlib==1.7.0
avro==1.12.1
cachetools==7.0.6
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
confluent-kafka==2.11.1
cryptography==47.0.0
Django==5.2.13
djangorestframework==3.17.1
enum-tools==0.13.0
fastavro==1.12.2
filelock==3.29.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.3.0
gast==0.7.0
google-pasta==0.2.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.12.0
idna==3.13
Jinja2==3.1.6
joserfc==1.6.4
jsonlines==4.0.0
jsonpath-ng==1.8.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.14.0
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
namex==0.1.0
networkx==3.6.1
numpy==2.4.4
opt_einsum==3.4.0
optree==0.19.0
orjson==3.11.8
packaging==26.2
propcache==0.4.1
protobuf==7.34.1
psycopg==3.3.3
psycopg-binary==3.3.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.20.0
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
quixstreams==3.23.6
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rich==14.3.4
rocksdict==0.3.29
rpds-py==0.30.0
safetensors==0.7.0
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.5
sympy==1.14.0
tensorflow==2.21.0
termcolor==3.3.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.6.2
twitchAPI==4.5.0
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wrapt==2.1.2
yarl==1.23.0
